requirements.txt
pandas
numpy
pyyaml
duckdb
pyarrow
requests
matplotlib
openai
